+++ lib/src/errors/error_creator.dart
@@ -3,7 +3,6 @@ import 'package:code_builder/code_builder.dart';
 import '../proto/contracts.pb.dart';
 import '../utils/rename_field.dart';
 
-// TODO: split into handlers?
 class ErrorCreator {
   const ErrorCreator();