+++ lib/src/statements/enum_handler.dart
@@ -1,5 +1,6 @@
 import 'package:code_builder/code_builder.dart';
 
+import '../attributes/attribute_creator.dart';
 import '../generator_database.dart';
 import '../types/type_creator.dart';
 import '../utils/rename_field.dart';
@@ -11,18 +12,21 @@ class EnumHandler extends StatementHandler {
   const EnumHandler(
     TypeCreator typeCreator,
     ValueCreator valueCreator,
+    AttributeCreator attributeCreator,
     GeneratorDatabase db,
-  ) : super(typeCreator, valueCreator, db);
+  ) : super(typeCreator, valueCreator, attributeCreator, db);
 
   @override
   Spec build(Statement statement) {
     final name = db.resolveName(statement.name);
 
-    // TODO: attributes
     return Enum(
       (b) => b
         ..name = name
-        ..docs.addAll(toDartdoc(statement.comment))
+        ..docs.addAll([
+          ...toDartdoc(statement.comment),
+          ...statement.attributes.map(attributeCreator.create),
+        ])
         ..values.addAll(
           statement.enum_11.members.map(
             (e) => EnumValue(